Update bin/test-integration's cloning of BOSH
This change allows you to do a file clone of BOSH into tmp/bosh in
the event that the network is too slow to do a shallow clone from
github. (File clones are shallow by default)

Additionally, it adds --depth 1 and --branch master to the clone if you
don't choose to use the file clone, which is quicker.