Setup networks after reading metadata settings and before hitting the registry
We want to support manual networking on OpenStack without dhcp configured, so
we need to support both dhcp and static IP reservation. So instead of
switching static/dhcp based on a stemcell flag, just set up all the networks
from the metadata service and then fetch the full settings from the registry.

As part of this we have one SetupNetworking for all networks instead of
SetupManualNetworking and SetupDhcpNetworks.

Lots of special cases removed from ubuntu and centos net managers.

We use MAC addresses to associate networks with interfaces. We're not
handling the case where there's only one network interface and the CPI hasn't
told us the MAC address for the network settings. We should just use the only
interface in that case.

[#84893840]

Signed-off-by: Colin OByrne <cobyrne@pivotal.io>