Pass a DefaultNetworkResolver to settings.Service instead of a platform
We were injecting a DefaultNetworkResolver into our net managers, which
platform delegated to, which we passed into settings.Service. Now we just pass
that dependency directly.

The DefaultNetworkResolver interface is in the settings package instead of the
net package to avoid an import cycle, which is maybe a little awkward.

[#84893840]

Signed-off-by: Colin OByrne <cobyrne@pivotal.io>