Fix cp_copier to work cross-platform
We were relying on "cp -R src/dir/ dest/dir" to not create
"dest/dir/dir" but on at least some linuxes it does. We now will only
copy files, and to continue supporting passing a directory as a filter
we test each filter to see if it's a directory and then convert it to a
"dir/**/*" filter before globbing.

Also fixes assertion in tarball_compressor_test to no longer rely on
the order of the output of "tar -t" which was inconsistent across
platforms.

[#89753100]

Signed-off-by: Tyler Schultz <tschultz@pivotal.io>