Actually wait for services to no longer be pending before stopping
Fixed the logic around detecting Pending. If services were Monitored and
Pending, they ended up in the monitoredServices array but not the
pendingServices array. Refactored some of the helper logic a bit.

Also fixed that we were ignoring errors on getting status while waiting
for services to no longer be pending

[#62354622](https://www.pivotaltracker.com/story/show/62354622)

Signed-off-by: Kam Leung <kleung@pivotal.io>