Support specifying physical interfaces to ignore
Currently, the agent configures every network interface that it finds to
obtain its configuration through DHCP, unless it receives from its
settings sources a network with a static configuration including a MAC
address.

While this works well for virtual machines where the network interfaces
are "virtual" and tipically attached to a network that provides DHCP,
that is not the case for baremetal deployments.  On the real hardware,
it is possible to find physical network interfaces were DHCPing would
always fail for some reason (e.g. cable unplugged, no DHCP server).
Disabling the problematic physical interfaces at hardware level is not a
generic solution, as some machines have dual port physical NICs where it
is not possible to disable one of the ports without disabling the whole
interface.

To prevent the Agent to DHCPing from a network interface where doing so
would always fail, provide a mechanism to specify a list of physical
interfaces that should be ignored by the Agent.

This commit also modifies the tests to exercise the new functionality.