Change run_scripts to take just a job script name and run for all jobs
This follows a conversation with Dmitriy where we decided it would be
best in the long run if the director doesn't know how the agent
organizes jobs on disk. All the director needs to know is that it wants
to run a script called "pre-start" that's bundled with each job.

[#101436550]
https://www.pivotaltracker.com/story/show/101436550

Signed-off-by: Jonathan Fuerth <jfuerth@pivotal.io>