Factored channel stuff out of the Script interface
Previously there was a real danger that a path through
GenericScript.Run() that does not write to the channel could be
introduced. Now this all funnels into a single line of code and that
particular danger is gone.

As a side effect, it's now also usable as a regular blocking function call.

Signed-off-by: Jamil Shamy <jshamy@pivotal.io>