remove real sleep inside WindowsNetManager
This was causing green tests to take 60 seconds in order to pass. We now inject
a Clock which gives us control over the flow of time. The tests now take 700ms
(due to using Consistently).

This required an update of the `github.com/pivotal-golang/clock` package which
meant that we needed to update the FakeClock we have vendored in this project.
The command we used to generate the fake was:

  counterfeiter -o agent/action/fakes/fake_clock.go ../../pivotal-golang/clock/clock.go Clock

[#119874373]

Signed-off-by: Cornelius Schumacher <cschum@suse.com>