make the `delete_arp_entries` action synchronous
We do not want this action to join the back of the queue because it may
affect the success of jobs ahead of it in the queue. For example, a
`pre-start` script may be failing due to having an invalid ARP cache and
that will block the queue preventing the fix from being applied until it
is too late.

[#119877249]

Signed-off-by: Christopher Brown <cbrown@pivotal.io>