Make sure that the created partition is available before continuing
We encountered issues with the partitioning while working on the
openSUSE stemcell.
The problem was that the new partitionins were not yet known to the
operating system so the subsequent steps failed.

To prevent issues like these we have added a `partprobe` call after
running parted to inform the operating system about the changes and
additionally an `udeavadm settle` call to ensure that the udev queue is
empty before continuing with the partition setup.