Windows: Implement GetConfiguredNetworkInterfaces
This resolves a bug where the Windows Agent would reset DNS settings
whenever the HTTPMetadataService was invoked.  The HTTPMetadataService
uses the GetConfiguredNetworkInterfaces method to determine if it needs
to perform a "minimal" network setup (ensureMinimalNetworkSetup), on
Windows that "minimal" setup resets the DNS settings.

Unfortunately there is no good was to determine if network interfaces
are configured, which is why this feature was not initially implemented,
to get around this SetupNetworking creates a file the first time it
succeeds and GetConfiguredNetworkInterfaces checks for the existence of
that file.

[#146886369](https://www.pivotaltracker.com/story/show/146886369)