Windows: refactor service management to be concurrent and not shell out
This commit refactors the Windows Job Supervisor to be perform its
operations concurrently and without shelling out.

The Start and Stop operations were always intended to act on all of the
services at once and without waiting for the requested action to
complete.  Unfortunately we recently discovered this was not the case
and this broke Stop Scripts that waited for all other jobs to stop
before stopping themselves - as the Supervisor would hanging trying to
stop it before advancing on to stop the other services.

The second and larger overvall change is that the Job Supervisor and
associated tests now use the 'golang.org/x/sys/windows/svc' package to
handle service operations instead of shelling out.  Shelling out to
PowerShell is extremely expensive (the service functions start at around
200ms on a fast VM, multiple seconds on a slow VM).  This is one of the
reason that Windows VMs were slow to show up when running `bosh
vms`.  Additionally, there was almost no error handling around the
commands we ran when shelling out - we were only checking if
PowerShell.exe ran correctly - not if any of the scripts threw an error,
which is not reflected in PowerShell's exit code (by default errors are
ignored).

[#148915345](https://www.pivotaltracker.com/story/show/148915345)