Windows: retry jobs that fail to start
If a job fails to start, attempt to restart it up to 5 times. Previously the
job supervisor did not check for errors during start and failing jobs where
automatically re-started by the service wrapper/service control manager.
We now check for errors during start, so to replicate the previous logic
while still being able to return a useful error we retry jobs that
failed to start and only return the error if none of the retries
succeeded.

[#150233685](https://www.pivotaltracker.com/story/show/150233685)