Fix goroutine leak in WindowsJobSupervisor
This commit fixes a goroutine leak in the WindowsJobSupervisor tests that
would occasionally lead to the runaway creation of goroutines.  This was
due to using Ifrit incorrectly.  As part of this commit we removed ifrit
as its original purpose was to make shutting down an http.Server easy,
which is no longer a concern with go 1.8.

[#150233685](https://www.pivotaltracker.com/story/show/150233685)

Signed-off-by: Natalie Arellano <narellano@pivotal.io>