Allow linux_platform_test to be run in parallel
This change updates the linux_platform_test so that it can be run in
parallel. The `SetupTmpDir` function was overriding the `TMPDIR`
environment variable which was causing Ginkgo to be unable to report for
all nodes. This change saves off the original `TMPDIR` in the BeforeEach
and then resets in the AfterEach.

Signed-off-by: Josh Russett <jrussett@pivotal.io>