Use parted when GPT partitions are detected
This change delegates the logic for partitioning persistent disks to a
specialized PersistentPartitioner. It will attempt to use `sfdisk` to
partition, but if it detects a GPT partition it will fall back to using
the parted partitioner.

This change also breaks apart the DeviceUtil interface. These two
objects were not being used interchangeably, and did not need to share
an interface.

[#153888031](https://www.pivotaltracker.com/story/show/153888031)

Signed-off-by: Josh Russett <jrussett@pivotal.io>