Kill and start monit daemon when on reload
This change reloads the monit daemon by running `sv kill monit` and `sv
start monit`. This is a faster way to reload the monit daemon then
`monit reload`. This is due to the fact that `monit reload` only occurs
after a heartbeat which occurs every 10 seconds. We should be safe to
run `sv kill monit` in both the apply and start actions because the
agent should have verified that all monit processes are stopped
previously.

[#156834041](https://www.pivotaltracker.com/story/show/156834041)

Co-authored-by: Difan Zhao <dzhao@pivotal.io>