Switch old micro_bosh path to use shared blobs dir
The old path was an artifact of microbosh and was only being used by
`create-env` environments. When deploying larger packages to a VM, it
could cause the system disk to be filled up.

Instead, switch to use the existing `/var/vcap/data/blobs` directory to
ensure we use the ephemeral disk. This also has the side-effect of no
longer respecting the `blobstore_path` which may be specified in
deployment manifests.

[#157071808](https://www.pivotaltracker.com/story/show/157071808)

Co-authored-by: Danny Berger <dberger@pivotal.io>