Downgrade go to 1.8.3
Changes in go1.9+ resulted in blocking reads when reading from
`/dev/random`. This has shown to be problematic when booting stemcells
in cloud infrastructures as the entropy is often very low on startup.
This change temporarily uses go1.8.3 to build the agent.

This change should be reverted when updating the agent to use go1.9+.

[#158403242](https://www.pivotaltracker.com/story/show/158403242)

Co-authored-by: James Myers <jmyers@pivotal.io>