Add feature flag to mount jobs directory on tmpfs
This change allows operators to mount their /var/vcap/data/jobs
directory (and /var/vcap/jobs, transitively) on a tmpfs to avoid the
contents being written to disk.

The configuration lives under the env.bosh instance group manifest
property:

  - name: redis
    ...
    env:
      bosh:
        job_dir:
	  tmpfs: true
	  tmpfs_size: 256m

This feature flag default to disabled and, if not set, the size of the
tmpfs defaults to 100 megabytes. The size is configurable in the
manifest to provide an interim escape-hatch if the default is too small
for a job which stores a large files in its job directory (this should
be uncommon).

The `FileBundle` change is required in order to avoid invalid
cross-device link errors when trying to rename a job directory. The
related tests are pended for the moment because they were very coupled
to the previous implementation. I'll update them if we're happy with
this approach. Please do not merge this commit!