integration: use single ssh connection for all tests
The overhead of running `vagrant ssh` on every single command being run
was outrageous. This change opens an SSH connection once at the start of
the tests and re-uses it thereafter. There are still usages of `vagrant
ssh` outside the standard `RunCommand` function which could be cleaned
up too but this is a reasonable start.

This causes the test suite on my workstation to go from 44 minutes to 4
minutes (11x speedup).