agentclient/http: fix retry format string
The format string was correct in the context of the informational line
but didn't have a placeholder for the error message added in the next
line.

Adding a '%s' to the first line would work but would confuse static
analysis tools which would detect it as a placeholder-to-argument
mismatch. Inlining everything results in a longer but overall simpler
line.