Simplify NicSettingsTemplate
We were running into a problem where after creating an HNS Network we
were getting two network adapters returned for the same MAC address.
This was then causing the next line of the powershell script to fail
since it was combining the two names.

We tried this change out with a 2019 stemcell. It fixed our problem and
seemed to otherwise work fine.

Signed-off-by: Leah Hanson <lhanson@pivotal.io>