Consider "loop" partition table as empty device
It appears that parted may show "loop" as the partition table for an
empty disk if it is not able to identify the partition table, but it
does appear to have some data on the device. This can theoretically
occur if a disk is not completely zeroed out. Either way, "loop" does
not appear to be a valid partition table, so should be safe to consider
as a new, empty disk.

We have seen this occur in some OpenStack environments, specifically
with ephemeral disks.

https://github.com/cloudfoundry/bosh/issues/2044
[#160688556](https://www.pivotaltracker.com/story/show/160688556)

Co-authored-by: David Ansari <david.ansari@sap.com>