decompress bundles directly into install path
In order to skip the disk-touching temporary directory interim step
inside the package and job appliers the FileBundle has been modified to
take a compressor and decompress the needed files from the bundle
directly into the install path.

This allowed us to unify the Windows and Unix install flows. We do not
expect the decompression to require a retry as it will not be modifying
existing files by renaming or deleting them. The uninstall step remains
separate.

[#161569370](https://www.pivotaltracker.com/story/show/161569370)

Co-authored-by: Christopher Brown <cbrown@pivotal.io>