allow explicitly choosing sfdisk for persistent disk
We previously added the ability to chose sfdisk explicitly for ephemeral
and root disk but missed persistent disks.

This adds the ability for one to use the persistent_disk_partitioner env
configuration with sfdisk.

This is a backport of commit from develop: e9ad9562014d1b3c90605dc36bcd4235705c52a2

[#161925673](https://www.pivotaltracker.com/story/show/161925673)

Co-authored-by: Belinda Liu <bliu@pivotal.io>