Retry extration with leading ./ on failure
Rendered job template archives take different forms depending on whether
or not they were delivered over NATS or via the blobstore. A leading ./
may be present. The old method of extracting everthing to an interim
temporary directory did not care about these specific pathing issues but
the new selective extraction does.

As a crude fix to get the pipeline green we retry with a leading ./ if
the extraction fails. We will add a proper fix which checks to see which
prefix to use without retrying unnecessarily tomorrow.

[#161569370](https://www.pivotaltracker.com/story/show/161569370)

Co-authored-by: Josh Russett <jrussett@pivotal.io>