integration/windows: uploaded blobs should be tgzs
The previous agent implementation didn't care whether or not job or
package blobs were compressed or not. However, it is expected that the
director will only ever upload compressed packages and so this test was
not representitive of reality.

I expect there will be other failures here but it doesn't seem to be
possible to easily `fly execute` the Windows integration tests.

[#161569370](https://www.pivotaltracker.com/story/show/161569370)