Remove job source from local blobstores after applying is successful
This change updates the bosh-agent to delete blobs associated with a
job's rendered templates after successfully applying and installing the
job. Previously, we never cleaned up these blobs and the growth of both
`/var/vcap/data/sensistive_blobs` and `/var/vcap/data/blobs` was
unbounded.

This change also updates the cascading blobstore to not attempt deletion
from the external blobstore. The external blobstore does not support
deletion, and the agent should not be responsible for managing blobs in
the director's external blobstore.

[finishes #162434271](https://www.pivotaltracker.com/story/show/162434271)

Co-authored-by: Belinda Liu <bliu@pivotal.io>