Agent can restart if tmpfs is enabled if VM did not reboot
Add a sentinel bootonce file to tmpfs
If the tmpfs bootonce file is still there, that means the VM did not
reboot

The tmpfs bootonce file is stored in the tmpfs settings directory even
if settings tmpfs is not enabled. This directory is always mounted even
if settings tmpfs is not enabled.

Rename RebootChecker -> StartManager and split up functionality for
readability

[#162990082](https://www.pivotaltracker.com/story/show/162990082)

Co-authored-by: Danny Berger <dberger@pivotal.io>