Fix path resolution when removing jobs
- 2012R2 has an issue when cleaning up jobs. The target directory of a
job is not a symlink as 2012R2 does not support ephemeral disks. This
led to filesystem.ReadAndFollowLink() returning an incomplete path
(without the drive letter).
- This led to it not matching the install and target directories and
jobs not getting cleaned up properly.
- The tests use a fake file system which does not extend itself to
capture this very specific Windows related filesystem behavior which is
not an issue on newer OS versions as they all have support for ephemeral
disks.
- By forcing the path comparison to resolve the absolute path of both
the install and the target directory, we can ensure that we compare two
complete paths and the clean-up happens as expected.

[#163153521]

Co-authored-by: Gab Satchi <gsatchithanantham@pivotal.io>