Bump WinSW to v2.2.0
It has been close to two years since we updated our service wrappers, so
we are updating them with the latest version. This *may* help resolve the
issue linked here; however, since these binaries are old, they needed to
be updating either way.

Further investigation will need to be done to
see if it resolves the linked story.

[#165039753](https://www.pivotaltracker.com/story/show/165039753)