Enhancement for encryption of persistent disk
This pull request includes the changes required to support usage and management of the encrypted disk in the bosh environment. Since the mount has to be done differently when the persistent disk is encrypted which needs encryption keys as well, so therefore it has to be handled externally (during deployment of encryption job) and not by the bosh-agent.
Seeing issues for scenarios like redeployment of vms or deletion/creation of vm having the same disk volume where bosh was trying to mount the already encrypted disk which now has fs type as crypto_LUKS instead of ext4/xfs.

Currently in bosh-agent code there is a method which returns cid to bosh director if the disk is mounted correctly. Since bosh is not able to mount the encrypted disk correctly in the above mentioned scenrio, therefore it is returning nil to the director, where AgentDiskOutOfSync check  compares the cid stored within the director to the one received by the agent which are now different in this case and therfore it raises exception saying "instance has invalid disks: agent reports 'nil' while director record shows vol-xxxx".

To fix this issue, the code change is done entirely on the bosh-agent side where it checks if the partition format type is crypto_LUKS through lsblk command, then don't try to mount the partition which will be done by the job later on and the variable isMountedExternally is set to true in this case.
One more filesystem type is added i.e crypto_LUKS apart from ext4 and xfs and now even at the time of formatting the disk this new fs type is also checked along with other fs types. If the disk is already formatted with crypto_LUKS type, it shouldn't be formatted again.

Co-authored-by: Maksim Yankovskiy <myankovskiy@zettaset.com>
Co-authored-by: Deepak Shetty <dshetty@zettaset.com>