Only remove files managed by bosh
The unit test fake filesystem currently behaves differently from
integration tests in a subtle way which was causing integration test
errors:

- In unit tests, the fake filesystem `Walk()` method would ignore the
containing directory so we did not ever remove `/etc/systemd/network`
- In integration tests, `Walk()` would yield that directory in addition
to files within it which would cause cleanup code to accidentally remove
the parent directory.

This change adds an exception such that we will never remove that
directory by accident. It does not yet change unit tests since that
change is more difficult to decide how best to make the change.

[#166361688]

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>