Don't check Content-Length when fetching blobs from signed urls
This change removes the assertion on Content-Length being equal to the
size of the file that we fetch from the blobstore when using signed
urls. This is due to the fact that the local webdav blobstore used in
most BOSH environments provides `-1` as the Content-Length. We have
confidence that the sha verification that occurs after is strong enough.

[#167583145](https://www.pivotaltracker.com/story/show/167583145)

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>