Prevent agents from restarting determinstically when NATS is not available
This change introduces exponential backoff and jitter to that initial
connection logic to NATS. It also increases the raw timeout and retries
when connecting.

This change also introduces an extended and randomized timeout when
publishing messages to the nats client. This prevents all of the agents
from exiting at the same time when the director is being deployed.

[#164934963](https://www.pivotaltracker.com/story/show/164934963)

Fixes #118

Co-authored-by: Charles Hansen <chansen@pivotal.io>