Add monkeypatch for Hash to work around vagrant-aws
See https://github.com/mitchellh/vagrant-aws/issues/566

Since we test the bosh-agent against linux and windows, we started
testing against go1.14 on windows and go1.13 on linux. I decided to
update the bosh-agent linux docker image to install go1.14.

The new docker image brought Ruby 2.5+ was installed and vagrant-aws needed to be updated.
I decided to keep going and update the plugin
(d6d7ba978f056792803fc52c22e0950848fdcb14).

This led to this error with vagrant-aws and Ruby's i18n. I could undo
all of this by pinning Ruby in our docker image so that none of these
dependencies are updated, but I think the risk of keeping this
monkeypatch around for a bit is fine.