Updates /etc/hosts template to have hostname first
According to the hosts(5) man page, the first name in the line of /etc/hosts
is the canonical hostname, and subsequent names are aliases for that
host. Experimentation agrees with the manual, so updating our /etc/hosts
generation to put the machine's hostname in the first position, rather
than the last, will cause `hostname -f` to return the host's hostname,
rather than "localhost".

[#178890404] BOSH-97 - 'hostname -f' command on bosh-deployed VMs returns 'localhost' instead of correct hostname