This moves /var/opt to /var/vcap/data/root_opt
Similar to what we do with /var/log, we've seen issues where
people are installing 3rd party software on stemcells using automation.

For example, the Microsoft OMI Agent.

This software gets installed into /var/opt and logs into /var/opt/omi/log.
This ends up filling up the root disk eventually and VMs need
to be recreated. This software has no configuration options to log
into /var/log (and it seems that logging into /var/opt/THING/log
may be the correct linux spot to log to)

While this is not the suggested way to install software
onto a stemcell, this type of software is often put there
by people not involved with Bosh lifecycles.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>