centos test symmetry: Added missing test: "when manual networks were configured with portable IP"
NOTE: Several of the expectations in the new/cloned test code are asymmetrical with the ubuntu test's code. One is probably because f3425689 (Write interfaces for ubuntu in alphabetic order) has not yet been migrated/replicated in the centos code. The other asymmetries seem to be due to implementation differences between the centos and ubuntu code which updates network interface config files.

See:
- commit 92e7502c81f2c56395abdd1daebd633ceb2cad5c: add test cases; fix syntax errors;
- commit f3425689692b9846d7ecd74d9c9d5ed59ac9b0fe: Support network alias
- commit 2c178e6c7f46b8964d8f97f5fc471788649b4687: Switch from ifupdown -> systemd-networkd
- commit 376fef8f36652e6c394f0bb06583d51d7b0060e5: Write interfaces for ubuntu in alphabetic order