Replace the `sfdisk -uM` command with a cross-platform compatible equivalent. (Fixes #265.)
This change adds explicit "multiplicative suffixes" in the sfdisk input to specify the units of the partition sizes (instead of the `-uM` command option, since `-uM` appears to be unsupported on many platforms and `-u` seems to be deprecated on many platforms). This change should produce identical partitioning behavior (vs the previous implementation), using an `sfdisk` input format which is more widely supported across linux platforms.

NOTE: Prior to this change, the `sfdisk -uM` command failed (on some stemcells), with agent log output like the following.

> ... ERROR - Failed with an error: Running command: 'sfdisk -uM /dev/sdc', stdout: '', stderr: 'sfdisk: unsupported unit 'M'

See:
- https://man7.org/linux/man-pages/man8/sfdisk.8.html#OPTIONS
- https://man7.org/linux/man-pages/man8/sfdisk.8.html#INPUT_FORMATS
- https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/8/html/considerations_in_adopting_rhel_8/file-systems-and-storage_considerations-in-adopting-rhel-8#removal-of-cylinder-head-sector-addressing-from-sfdisk-and-cfdisk_file-systems-and-storage