Add setup nats fw implementation linux.
this commit contains the linux counterpart for https://github.com/cloudfoundry/bosh-agent/pull/259

This adds the setup firewall implementation based on containerds
cgroups and the go-iptables package. It is a replacement for the
stemcell based approach that didn't work reliably.

On VM creation bosh provides the Agent on the VM with creds
to connect to Nats. Worklaoads that live on the vm could
extract these credentials and use them to connect to nats and
extract further credentials that are being sent over nats once
the rendering of jobs takes place.

It returns early if the mbus url is an empty string:
- e.g. http_metadata_service calls SetupNetworking
before we can possibly have the mbus url because it is
responsible for extracting the agent settings.json from the
metadata endpoint.

or if it mbus url starts with https://:
- on a create-env we will see the mbus URL start with https as this
is where the cli posts the data via a tunnel. In that case we do
not want a firewall rule because the ssh tunnel the bosh-cli uses
will have the traffic blocked (since the ssh session wouldn't be
part of the cgroup) and therefore the create-env will fail because
the cli cannot send msgs.

the rest of the code
- sets up the iptables rules to block outgoing traffic to the mbus
  endpoint found in agent settings.json based on cgroup membership
- sets up the cgroup and adds it's own PID to the created cgroup.

Additional changes:
- updated the RunCommand to actually include stderr if the command
  on the target host fails
- added unit tests for return early behaviour
- added integration test for the created gw rules setup
- updated windows implementation to also return early
Tracker Stories #180698114 #181114075