Add V6 iptables support
Following the merge of the Nats FW feature, sStemcells build with the agent will
not be able to successfully deploy in a V6 setup. The reason/background is that
the Iptables code is trying to add V6 addresses via the Iptables v4 API.

Add logic to figure out if the actual IP behind the host in `mbus_string`
is an IPv4 vs IPv6 address and handle instantiation of Iptables lib accordingly

Monit runs on V4 locally, even on a V6 VM thus the created rule is always V4.