Monit runs on V4 locally, even on a V6 VM
b700675bbe63c7cc389e3edc2a63f102a7291567 Added handling for
ipv6 based iptables. Unfortunately it was trying to be too
smart. The resulting stemcell still broke timing out pinging
to agent but the logs on the VM showed a different error.

Previously it would fail on adding a rule with a v6 address
to a v4 table. Now the logs indicated the bosh was sending
status checks for monit and the agent timed out as the agents
cgroup only allowed a V6 loopback target. This switches the
behaviour to always create the monit rule in the v4 table and
conditionally create the nats rule in the appropriate table