resolv.conf: centos: Do not overwrite /etc/resolv.conf in SetupNetworking (as it is controlled by Network Manager).
NOTE: The change to NOT write to `resolv.conf` (for centos) was originally made in 63548d43. Thinking the missing/asymmetric code in centosNetManager was a bug, I had tried re-adding the missing/asymmetric functionality to centosNetManager, but discovered during testing that the 'missing' functionality _needed_ to be missing (due to differences between centos vs ubuntu networking). This commit adds explicit code comments at the locations where the centosNetManager code is intentionally asymmetric (vs ubuntuNetManager), and adds tests which explicitly verify the original (and correct) behavior introduced in 63548d43, to prevent anyone else from repeating my mistake of restoring symmetry (at least in this particular context) between the centos and ubuntu code/functionality.

See:
- commit 63548d43c69180b761d96b8e42a699e0762779e2: On Centos, write DNS config in device sysconfig
- https://ma.ttias.be/centos-7-networkmanager-keeps-overwriting-etcresolv-conf/
- https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/8/html/configuring_and_managing_networking/manually-configuring-the-etc-resolv-conf-file_configuring-and-managing-networking
- https://wiseindy.com/blog/linux/how-to-set-dns-in-centos-rhel-7-prevent-network-manager-from-overwriting-etc-resolv-conf/