Add logging around nats disconnects
We are seeing intermittent unresponsive agents in following scenario:

- upgrade bosh-director
- right after upgrade is finished, start recreating a deployment

In some situations there seems to be a race condition in between

1. the old director going down
2. Agent on the VM retries sending heartbeets
3. new director comes up and get's deploy command issued. Sends ping to agent
4. ???? Currently unclear why at this point the agent has not yet reconnected to nats
5. director doesn't get a response and the deploy command fails.

the issue has been solved in CI by introducing a 3 minute wait after the
upgrade of the director but we would like to get to the root cause of
this. It seems that if the agent DID NOT crash and restart within BOSHs downtime,
the connection is somehow in a bad state. Cunnie suggested it that the
arp cache could be the reason the connection is not successfully re-
established after the director comes back up.

Use nats client handler functions to add additional logging and an arp
flush on connection issues.