Upgrade nats-io/nats-server/v2 for CVEs
This *seems* to be an indirect dependency of the nats-io/nats.go
package. It's referenced in a `go_test.mod` file for that package.
Dependabot has detected a number of vulnerabilities for v2.0.0 of
nats-server/v2.

However:
1. There are no references in nats-io/nats.go beyond go_test.mod
2. There are no references in the agent
3. The code for nats-io/nats-server is not even vendored.

None of the standard commands to upgrade dependencies affected this
package. The only thing that seemed to work is either removing it from
go.mod and running `go mod tidy` to add it back, or to do an explicit
`go get github.com/nats-io/nats-server/v2`.

[#181936278] Auto-bump test dependencies in bosh-agent